Extrait la méthode de formattage du numéro de téléphone
